Moved level3 kernel files to a subfolder
